tests/test_predictor.py
import unittest
from predictor import get_signal

# Sample OHLCV data (replace with realistic data if needed for your actual strategy tests)
# [timestamp, open, high, low, close, volume]
SAMPLE_OHLCV_DATA = [
    [1678886400000, 0.38, 0.39, 0.37, 0.385, 100000],
    [1678886460000, 0.385, 0.395, 0.38, 0.39, 120000],
    [1678886520000, 0.39, 0.392, 0.388, 0.391, 80000],
    # Add more data points as required by your strategy
]

class TestPredictor(unittest.TestCase):

    def test_get_signal_placeholder(self):
        """Test the placeholder behavior which should always return 'NONE'"""
        symbol = 'XRP/USDT:USDT'
        timeframe = '1m'
        signal = get_signal(symbol, timeframe, SAMPLE_OHLCV_DATA)
        self.assertEqual(signal, 'NONE', "Placeholder should return 'NONE'")

    def test_get_signal_no_data(self):
        """Test behavior when no OHLCV data is provided."""
        symbol = 'XRP/USDT:USDT'
        timeframe = '1m'
        signal = get_signal(symbol, timeframe, [])
        self.assertEqual(signal, 'NONE', "Should return 'NONE' when data is empty")

    # --- Add more tests below when you implement your strategy --- 
    # Example: Test for a LONG signal under specific conditions
    # def test_get_signal_long_condition(self):
    #     specific_ohlcv = [...] # Create data that SHOULD trigger a LONG
    #     signal = get_signal('XRP/USDT:USDT', '1m', specific_ohlcv)
    #     self.assertEqual(signal, 'LONG', "Should return 'LONG' for specific condition")

    # Example: Test for a SHORT signal under specific conditions
    # def test_get_signal_short_condition(self):
    #     specific_ohlcv = [...] # Create data that SHOULD trigger a SHORT
    #     signal = get_signal('XRP/USDT:USDT', '1m', specific_ohlcv)
    #     self.assertEqual(signal, 'SHORT', "Should return 'SHORT' for specific condition")

if __name__ == '__main__':
    unittest.main() 